Add doc comments to new classes and members
