+++ lib/messages.dart
@@ -34,9 +34,9 @@ part 'src/messages/requests/create_checkout_session.dart';
 part 'src/messages/requests/create_customer.dart';
 part 'src/messages/requests/create_customer_balance_transaction_request.dart';
 part 'src/messages/requests/create_discount.dart';
+part 'src/messages/requests/create_invoice.dart';
 part 'src/messages/requests/create_invoice_item.dart';
 part 'src/messages/requests/create_invoice_preview_request.dart';
-part 'src/messages/requests/create_invoice_request.dart';
 part 'src/messages/requests/create_payment_intent.dart';
 part 'src/messages/requests/create_portal_session.dart';
 part 'src/messages/requests/create_price.dart';
